refactor: separate release and WHY responsibilities
